backend/pyproject.toml
[project]
name = "easystash"
version = "0.1.0"
description = ""
authors = [
    {name = "TheMrKan",email = "mrkan2017@gmail.com"}
]

requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-filter (>=25.1,<26.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "easystash", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
